compile unit: FieldJ and stage compute current
